requirements.txt
httpx>=0.27
pydantic>=2.6
loguru>=0.7
python-dotenv>=1.0
tenacity>=8.2
PyYAML>=6.0
openai>=1.37

